Get feature to players can request to join a team
Current members of the team can respond to the request as well.
