+++ app/views/player_information.py
@@ -4,7 +4,7 @@ from flask import render_template
 from flask_login import login_required, current_user
 from app import wjl_app
 from app.views.helper import get_base_data
-from app.model import Field, Team
+from app.model import Field, Team, LeagueRequest
 from app.errors import NotFoundException
 
 
@@ -27,7 +27,12 @@ def team(team_id: int):
     team = Team.query.get(team_id)
     if team is None:
         raise NotFoundException(f"Sorry, team not found - {team_id}")
+    team_requests = (LeagueRequest.query
+                     .filter(LeagueRequest.email == current_user.email)
+                     .filter(LeagueRequest.team_id == team_id)
+                     .filter(LeagueRequest.pending == True).count())
     return render_template("team.html",
                            team=team.json(),
                            on_team=current_user in team.players,
+                           made_request=team_requests > 0,
                            base_data=get_base_data())
